isolate extractor change and proper error handling
