+++ baseapp/baseapp.go
@@ -623,7 +623,6 @@ func (app *BaseApp) CheckTx(txBytes []byte) (res abci.ResponseCheckTx) {
 func (app *BaseApp) DeliverTx(txBytes []byte) (res abci.ResponseDeliverTx) {
 	var result sdk.Result
 	tx, err := app.txDecoder(txBytes)
-	txHash := fmt.Sprintf("%X", tmhash.Sum(txBytes))
 	if err != nil {
 		result = err.Result()
 	} else {
@@ -640,72 +639,10 @@ func (app *BaseApp) DeliverTx(txBytes []byte) (res abci.ResponseDeliverTx) {
 		Tags:      result.Tags,
 	}
 
-	ctx := app.getContextForTx(runTxModeDeliver, txBytes)
-
-	sdktx, _ := tx.(auth.StdTx)
-	jsonTags, _ := codec.Cdc.MarshalJSON(sdk.TagsToStringTags(result.Tags))
-	jsonMsgs := MsgsToString(sdktx.GetMsgs())
-	jsonFee, _ := codec.Cdc.MarshalJSON(sdktx.Fee)
-	jsonMemo, _ := codec.Cdc.MarshalJSON(sdktx.GetMemo())
-
-	for idx, msg := range sdktx.GetMsgs() {
-		f, _ := os.OpenFile(fmt.Sprintf("./extract/progress/messages.%d.%s", ctx.BlockHeight(), ctx.ChainID()), os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
-		msgString, _ := codec.Cdc.MarshalJSON(msg)
-		f.WriteString(fmt.Sprintf("%s,%d,%s,$%s$,%s,%s\n",
-			txHash,
-			idx,
-			msg.Type(),
-			strings.ReplaceAll(string(msgString), "$", "\\$"),
-			ctx.BlockHeader().Time.Format("2006-01-02 15:04:05"),
-			ctx.ChainID()))
-		f.Close()
-		extractAddresses(msg, string(txHash), ctx.BlockHeight(), idx, ctx.ChainID())
-	}
-
-	f, _ := os.OpenFile(fmt.Sprintf("./extract/progress/txs.%d.%s", ctx.BlockHeight(), ctx.ChainID()), os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
-	f.WriteString(fmt.Sprintf("%s,%d,%d,%d,%d,$%s$,$%s$,$%s$,$%s$,$%s$,%s,%s\n",
-		txHash,
-		ctx.BlockHeight(),
-		uint32(result.Code),
-		int64(result.GasWanted),
-		int64(result.GasUsed),
-		strings.ReplaceAll(result.Log, "$", "\\$"),
-		strings.ReplaceAll(string(jsonMemo), "$", "\\$"),
-		strings.ReplaceAll(string(jsonFee), "$", "\\$"),
-		strings.ReplaceAll(string(jsonTags), "$", "\\$"),
-		strings.ReplaceAll(string(jsonMsgs), "$", "\\$"),
-		ctx.BlockHeader().Time.Format("2006-01-02 15:04:05"),
-		ctx.ChainID()))
-	f.Close()
+	recordTxData(app, txBytes, tx, response)
 
 	return response
 
-}
-
-func getAddresses(msg sdk.Msg) []sdk.Address {
-	var addrs []sdk.Address
-	m := BasicMsgStruct{}
-	_ = json.Unmarshal(msg.GetSignBytes(), &m)
-	ref := reflect.ValueOf(&m.Value).Elem()
-	for i := 0; i < ref.NumField(); i++ {
-		addr := ref.Field(i).Interface()
-		sdkAddr, ok := addr.(sdk.Address) // cast to address interface so we have access to the String() method, which bech32ifies the address
-		if ok && !sdkAddr.Empty() {
-			addrs = append(addrs, sdkAddr)
-		}
-	}
-	return addrs
-}
-
-func extractAddresses(msg sdk.Msg, hash string, height int64, idx int, chainid string) {
-	addrs := getAddresses(msg)
-	if len(addrs) != 0 {
-		f, _ := os.OpenFile(fmt.Sprintf("./extract/progress/addresses.%d.%s", height, chainid), os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
-		for _, addr := range addrs {
-			f.WriteString(fmt.Sprintf("%s,%d,%s,%s\n", hash, idx, addr.String(), chainid))
-		}
-		f.Close()
-	}
 }
 
 type (
@@ -985,45 +922,6 @@ func (app *BaseApp) runTx(mode runTxMode, txBytes []byte, tx sdk.Tx) (result sdk
 	}
 
 	return
-}
-
-func copyFile(destination string, source string) error {
-	sourceFile, err := os.OpenFile(source, os.O_RDONLY, 0644)
-	if err != nil {
-		return err
-	}
-	defer sourceFile.Close()
-
-	destinationFile, err := os.OpenFile(destination, os.O_WRONLY | os.O_CREATE | os.O_APPEND, 0644)
-	if err != nil {
-		return err
-	}
-	defer destinationFile.Close()
-
-	_, err = io.Copy(destinationFile, sourceFile)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func commitUncheckedFiles(ctx sdk.Context) {
-	for _, key := range []string{"delegations", "unbond", "balance", "rewards"} {
-		err := copyFile(fmt.Sprintf("./extract/progress/%s.%d.%s", key, ctx.BlockHeight(), ctx.ChainID()), fmt.Sprintf("./extract/unchecked/%s.%d.%s", key, ctx.BlockHeight(), ctx.ChainID()))
-		if err != nil {
-			fmt.Errorf("error: %v while commiting unchecked file\n", err)
-			continue
-		}
-		// No need for the file now
-		os.Remove(fmt.Sprintf("./extract/unchecked/%s.%d.%s", key, ctx.BlockHeight(), ctx.ChainID()))
-	}
-}
-
-func deleteUncheckedFiles(ctx sdk.Context) {
-	for _, key := range []string{"delegations", "unbond", "balance", "rewards"} {
-		os.Remove(fmt.Sprintf("./extract/unchecked/%s.%d.%s", key, ctx.BlockHeight(), ctx.ChainID()))
-	}
 }
 
 // EndBlock implements the ABCI interface.
